Desktop/khareetaty-ai-mvp/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=1.5.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
prophet>=1.1.0
scikit-learn>=1.1.0
geopy>=2.3.0
streamlit>=1.20.0
plotly>=5.10.0
pytz>=2022.0
requests>=2.28.0
twilio>=8.10.0
email-validator>=1.3.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
APScheduler>=3.10.0
pyjwt>=2.8.0
bcrypt>=4.0.0
passlib>=1.7.4
shapely>=2.0.0
geojson>=3.0.0
rtree>=1.0.0